+++ tests/testthat/test-autoplot.curve_params.R
@@ -5,7 +5,7 @@ test_that(
     curve <-
       serocalculator_example("example_curve_params.csv") |>
       read.csv() |>
-      as_curve_params() |>
+      as_sr_params() |>
       filter(antigen_iso %in% c("HlyE_IgA", "HlyE_IgG")) |>
       autoplot()